Fix several MSVC compiler warning/errors

+++ vpx_ports/mem.h
@@ -44,7 +44,7 @@
 
 /* Shift down with rounding for use when n >= 0 */
 #define ROUNDZ_POWER_OF_TWO(value, n) \
-    ((n) ? (((value) + (1 << ((n) - 1))) >> (n)) : (value))
+    ((n) ? (((value) + (int64_t)(1 << ((n) - 1))) >> (n)) : (value))
 
 #define ALIGN_POWER_OF_TWO(value, n) \
     (((value) + ((1 << (n)) - 1)) & ~((1 << (n)) - 1))
